Refactor class names and paths
